feat: call SetupStream and SetupMetrics from nginx configuration (#105)
* feat: remove stream mode and metrics setup

Removed the unused stream mode setup and metrics setup from the ~allowIp~ and ~Allow~ functions to clean up the code.

* feat(lib): add get_mode function to retrieve runtime mode

Added a new function ~csmod.get_mode~ to retrieve the runtime mode from the configuration. This function returns the value of the ~MODE~ key from the ~runtime.conf~ table.

* feat(crowdsec): expose SetupMetrics and SetupStream functions

Expose the SetupMetrics and SetupStream functions to make them accessible outside the module.

* fix: correct function reference in timer setup

Corrected the function reference in the timer setup from ~Setup_metrics~ to ~csmod.SetupMetrics~ to ensure proper execution of the timer.

* feat: add init_worker_by_lua_block for stream mode and metrics initialization

+ Added ~init_worker_by_lua_block~ to initialize stream mode and metrics for workers.
+ Ensured that stream mode is set up for workers when mode is "stream".
+ Metrics initialization is performed only by worker 0.

* fix: correct debug cache order in test file

Ensure the debug cache entries are in the correct order in `t/05stream_and_ban.t`.